mnmcp-core/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mnmcp-core"
version = "1.0.0"
description = "MnMCP - Minecraft and MiniWorld Cross-Platform Protocol Engine"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "websockets>=12.0",
    "cryptography>=41.0.0",
    "pyyaml>=6.0.0",
    "protobuf>=4.25.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0","pytest-asyncio>=0.21.0","black>=23.7.0","ruff>=0.1.0"]
gui = ["customtkinter>=5.2.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
